+++ NEWS.md
@@ -7,6 +7,8 @@
 * `libxml2_version()` now returns a semantic version rather than alphanumeric
   version, so "2.9.10" > "2.9.9" (#277)
 
+* `as_xml_document()` now handles cases with text nodes trailing normal nodes (#274).
+
 # xml2 1.2.2
 
 * Export S4 classes with documentation, so they can be used in other packages